+++ packages/pagebuilder/lib/ContentTypes/Banner/configAggregator.js
@@ -5,11 +5,9 @@ import {
     getPadding,
     getTextAlign,
     getCssClasses,
-    getIsHidden
+    getIsHidden,
+    injectStoreCodeHref
 } from '../../utils';
-
-import { BrowserPersistence } from '@magento/peregrine/lib/util';
-const storage = new BrowserPersistence();
 
 /**
  * Determine the button type based on class
@@ -43,17 +41,7 @@ export default (node, props) => {
         minHeightPaddingElement = overlayElement;
     }
 
-    if (process.env.USE_STORE_CODE_IN_URL === 'true') {
-        const storeViewCode =
-            storage.getItem('store_view_code') || STORE_VIEW_CODE;
-
-        // For each link update the href to include the active store code.
-        node.querySelectorAll('a').forEach(oldNode => {
-            const newNode = oldNode.cloneNode(true);
-            newNode.href = `/${storeViewCode}${oldNode.getAttribute('href')}`;
-            oldNode.parentNode.replaceChild(newNode, oldNode);
-        });
-    }
+    node = injectStoreCodeHref(node);
 
     return {
         minHeight: minHeightPaddingElement.style.minHeight || null,
